reverted unrelated changes, moved thumbnails.py to utilities repo
