+++ src/routes/articles/[dirslug]/[articleslug]/+layout.svelte
@@ -16,7 +16,7 @@
   import { browser } from '$app/environment';
 
   import { page } from '$app/stores';
-  import PageMeta from '$lib/linkpreview/PageMeta.svelte';
+  import PageMeta from '$lib/metadata/PageMeta.svelte';
 
   let isOpen = false;
   export let anchors = [];
